requirements.txt
pyyaml>=6.0.1
cryptography>=41.0.0
loguru>=0.7.0